Moved MTpipeline/ephem to mtpipline/ephem. Removed duplicate test_jpl2db.py from ephem. Updated import statements run_trim_test.py and test_build_finders_table.py, so that they will hopefully work.
