+++ mtpipeline/ephem/build_master_images_table.py
